test: keep launch unit cases pure
